requirements.txt
streamlit==1.48.1
python-dotenv
llama-index
llama-index-llms-google-genai
llama-index-embeddings-google-genai

